Improvements: 1. data layer recycled: divided on data and datasource layers 2. add converter for fetch history 3. presentation layer recycled: distributed by features packages
