+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
         <activity
-                android:name="com.readyfo.yaclone.presentation.ui.activities.SplashScreenActivity"
+                android:name="com.readyfo.yaclone.presentation.activities.SplashScreenActivity"
                 android:theme="@style/AppThemeLauncher">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -20,7 +20,7 @@
             </intent-filter>
         </activity>
         <activity
-                android:name="com.readyfo.yaclone.presentation.ui.activities.MainActivity"
+                android:name="com.readyfo.yaclone.presentation.activities.MainActivity"
                 android:label="@string/app_name"
                 android:launchMode="singleTop"
                 android:windowSoftInputMode="adjustPan">
