+++ app/src/main/java/com/readyfo/yaclone/data/FetchHistoryRepositoryImpl.kt
@@ -1,0 +1,12 @@
+package com.readyfo.yaclone.data
+
+import com.readyfo.yaclone.data.datasource.FetchHistoryDataSource
+import com.readyfo.yaclone.domain.repositories.FetchHistoryRepository
+import javax.inject.Inject
+
+class FetchHistoryRepositoryImpl @Inject constructor(
+     val fetchHistoryDataSource: FetchHistoryDataSource
+) : FetchHistoryRepository {
+
+    override fun fetchHistory() = fetchHistoryDataSource.fetchHistory()
+}
