+++ app/src/main/java/com/readyfo/yaclone/data/datasource/FetchHistoryDataSource.kt
@@ -1,8 +1,8 @@
-package com.readyfo.yaclone.domain.usecases
+package com.readyfo.yaclone.data.datasource
 
 import com.readyfo.yaclone.domain.models.HistoryModel
 import io.reactivex.Observable
 
-interface FetchHistoryUseCase {
+interface FetchHistoryDataSource {
     fun fetchHistory(): Observable<MutableList<HistoryModel>>
 }
