+++ app/src/main/java/com/readyfo/yaclone/di/AppComponent.kt
@@ -1,7 +1,7 @@
 package com.readyfo.yaclone.di
 
 import com.readyfo.yaclone.di.modules.AppModule
-import com.readyfo.yaclone.presentation.ui.fragments.HistoryFragment
+import com.readyfo.yaclone.presentation.history.HistoryFragment
 import dagger.Component
 
 @Component(modules = [AppModule::class])
