+++ app/src/main/java/com/readyfo/yaclone/domain/models/HistoryModel.kt
@@ -1,0 +1,10 @@
+package com.readyfo.yaclone.domain.models
+
+data class HistoryModel(
+    val id: Int,
+    val textFrom: String,
+    val textTo: String,
+    val languageFrom: String,
+    val languageTo: String,
+    val favorite: Boolean
+)
