+++ app/src/main/java/com/readyfo/yaclone/presentation/activities/MainActivity.kt
@@ -1,0 +1,30 @@
+package com.readyfo.yaclone.presentation.activities
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.readyfo.yaclone.R
+
+
+class MainActivity : AppCompatActivity() {
+    private lateinit var bottomNavView: BottomNavigationView
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        setupNavigation()
+    }
+
+    override fun onSupportNavigateUp(): Boolean {
+        return findNavController(R.id.nav_host_fragment).navigateUp()
+    }
+
+    private fun setupNavigation() {
+        bottomNavView = findViewById(R.id.nav_view)
+        val navController = findNavController(R.id.nav_host_fragment)
+        bottomNavView.setupWithNavController(navController)
+    }
+}
