+++ app/src/main/java/com/readyfo/yaclone/presentation/activities/SplashScreenActivity.kt
@@ -1,0 +1,18 @@
+package com.readyfo.yaclone.presentation.activities
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.readyfo.yaclone.R
+
+class SplashScreenActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash_screen)
+
+        val intent = Intent(this, MainActivity::class.java)
+        startActivity(intent)
+        finish()
+    }
+}
