--- app/src/main/java/com/readyfo/yaclone/presentation/presenters/TranslatorPresenter.kt
@@ -1,9 +1,0 @@
-package com.readyfo.yaclone.presentation.presenters
-
-import com.readyfo.yaclone.presentation.views.TranslatorView
-import moxy.InjectViewState
-import moxy.MvpPresenter
-
-@InjectViewState
-class TranslatorPresenter: MvpPresenter<TranslatorView>() {
-}
