+++ app/src/main/java/com/readyfo/yaclone/presentation/translator/TranslatorPresenter.kt
@@ -1,0 +1,8 @@
+package com.readyfo.yaclone.presentation.translator
+
+import moxy.InjectViewState
+import moxy.MvpPresenter
+
+@InjectViewState
+class TranslatorPresenter: MvpPresenter<TranslatorView>() {
+}
