+++ app/src/main/java/com/readyfo/yaclone/presentation/translator/TranslatorView.kt
@@ -1,0 +1,6 @@
+package com.readyfo.yaclone.presentation.translator
+
+import moxy.MvpView
+
+interface TranslatorView: MvpView {
+}
