--- app/src/main/java/com/readyfo/yaclone/presentation/ui/fragments/ChoseLangFragment.kt
@@ -1,20 +1,0 @@
-package com.readyfo.yaclone.presentation.ui.fragments
-
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.readyfo.yaclone.R
-
-class ChoseLangFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_chose_lang, container, false)
-    }
-}
