--- app/src/main/java/com/readyfo/yaclone/presentation/ui/fragments/TranslatorFragment.kt
@@ -1,27 +1,0 @@
-package com.readyfo.yaclone.presentation.ui.fragments
-
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.readyfo.yaclone.R
-import moxy.MvpAppCompatFragment
-
-
-class TranslatorFragment : MvpAppCompatFragment(){
-//    @InjectPresenter
-//    var translatorPresenter: TranslatorPresenter? = null
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_translator, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
