--- app/src/main/java/com/readyfo/yaclone/presentation/views/TranslatorView.kt
@@ -1,6 +1,0 @@
-package com.readyfo.yaclone.presentation.views
-
-import moxy.MvpView
-
-interface TranslatorView: MvpView {
-}
