+++ app/src/main/res/layout-land/fragment_translator.xml
@@ -5,7 +5,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/colorPrimary"
-        tools:context=".presentation.ui.fragments.TranslatorFragment">
+        tools:context=".presentation.translator.TranslatorFragment">
     <LinearLayout
             android:id="@+id/choseLangContainer"
             android:layout_width="wrap_content"
