+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
         android:id="@+id/container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.activities.MainActivity">
+        tools:context=".presentation.activities.MainActivity">
 
     <fragment
             android:id="@+id/nav_host_fragment"
