+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.activities.SplashScreenActivity">
+        tools:context=".presentation.activities.SplashScreenActivity">
     <ImageView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
