+++ app/src/main/res/layout/fragment_chose_lang.xml
@@ -4,7 +4,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.fragments.ChoseLangFragment">
+        tools:context=".presentation.choselang.ChoseLangFragment">
 
     <include
             android:id="@+id/toolbar"
