+++ app/src/main/res/layout/fragment_history.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.fragments.HistoryFragment">
+        tools:context=".presentation.history.HistoryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerViewHistory"
